Data grabbing seperated from historical page
